Update tokenizer.js
maybe this is it? dont want it as a punct? 

i dont see ?. in here, theres  a special case for that, so maybe thats needed for ?? in the same spot??? EXECPT you can have a space before/after the whole token for ?? and im guessing the done do that for .? n thats why theres the special case